requirements.txt
Flask
flask_SQLAlchemy
flask_restful
psycopg2
